Clean up transcription code syntax; rename some files; move some imports around
