--- src/py_server/background_removal.py
@@ -1,2 +1,0 @@
-def remove_background():
-    raise NotImplementedError
