--- src/sharedTypes/Transcription.ts
@@ -1,9 +1,0 @@
-import Word from './Word';
-
-interface Transcription {
-  confidence: number;
-
-  words: Word[];
-}
-
-export default Transcription;
