--- src/sharedTypes/Word.ts
@@ -1,9 +1,0 @@
-interface Word {
-  word: string;
-
-  start_time: number;
-
-  duration: number;
-}
-
-export default Word;
